+++ src/Bpp/Seq/Container/AbstractSequenceContainer.cpp
@@ -48,7 +48,7 @@ using namespace std;
 void AbstractSequenceContainer::setComments(const string& name, const Comments& comments)
 throw (SequenceNotFoundException)
 { 
-	unsigned int pos = getSequencePosition(name);
+	size_t pos = getSequencePosition(name);
 	setComments(pos, comments);
 }
