+++ src/Bpp/Seq/DistanceMatrix.cpp
@@ -42,9 +42,9 @@ using namespace std;
 
 using namespace bpp;
 
-unsigned int DistanceMatrix::getNameIndex(const std::string& name) const throw (Exception)
+size_t DistanceMatrix::getNameIndex(const std::string& name) const throw (Exception)
 {
-  for (unsigned int i = 0; i < names_.size(); ++i)
+  for (size_t i = 0; i < names_.size(); ++i)
   {
     if (names_[i] == name) return i;
   }
